chore: remove Insights feature; extract card-scrubber module
The Insights builder/viewer ended up not fitting the program's direction
(focus is now Studio, Transcripts, Screenspace) so delete it cleanly: 10
source files, --insights flag, /insights/ blueprint, sprite-sheet backend,
data_export insights table, related tokens and docs. Before deleting,
extract the hover-to-scrub + audio-snippet + waveform overlay into a
reusable assets/web/card-scrubber.{js,css} module exposing
window.clipgenCardScrubber.attach/detachAll/stopAll, ready for a future
toggleable "fancy cards" mode in Studio. MARK_CATEGORIES.insight (the
unrelated transcript mark) is preserved.
