+++ agents/ARCHITECTURE.md
@@ -19,13 +19,11 @@
 | [config.py](config.py) | Global constants and settings (version, headers, limits, commands) |
 | [google_api.py](google_api.py) | Google Sheets auth, worksheet selection by priority, spreadsheet listing/search |
 | [excel_io.py](excel_io.py) | Excel adapter: `ExcelSheetAdapter` mimics gspread Worksheet interface for local .xlsx |
-| [server.py](server.py) | Combined Flask server for Studio + Insights + Screenspace; registers blueprints per active mode, `start_combined_server()` handles all three on one port |
+| [server.py](server.py) | Combined Flask server for Studio + Screenspace + Transcripts; registers blueprints per active mode, `start_combined_server()` handles all three on one port |
 | [screenspace.py](screenspace.py) | Screenspace analysis engine: image analysis primitives, ten analysis tools (color, change, similarity, text, numbers, timelapse, template, flow, scene, inactivity) plus a `multitool` chaining mode, background task queue worker (`ScreenspaceWorker`), manifest persistence |
 | [screenspace_server.py](screenspace_server.py) | Screenspace Flask REST API: region CRUD, video frame extraction, task queue management, results retrieval |
-| [insights.py](insights.py) | Insights data model: CRUD operations for insight records, insights manifest read/write |
-| [insights_server.py](insights_server.py) | Insights Flask REST API: insight CRUD, artifact browsing, sprite sheet generation, viewer export |
-| [data_export.py](data_export.py) | Analysis-ready JSON+CSV export from Screenspace, Insights, and Transcripts manifests; powers `--export` CLI flag and `/screenspace/api/export/events` endpoint |
-| [assets/web/](assets/web/) | Static HTML/JS/CSS templates: timeline viewer (`viewer.html/js/css`), gallery (`gallery.html/js/css`), studio (`studio.html/js/css`), insights (`insights-builder.html/js/css`), insights viewer (`insights-viewer.html/js/css`), screenspace (`screenspace.html/js/css`). Shared utilities and constants live in `utils.js` (loaded before each page's main script) |
+| [data_export.py](data_export.py) | Analysis-ready JSON+CSV export from Screenspace and Transcripts manifests; powers `--export` CLI flag and `/screenspace/api/export/events` endpoint |
+| [assets/web/](assets/web/) | Static HTML/JS/CSS templates: timeline viewer (`viewer.html/js/css`), gallery (`gallery.html/js/css`), studio (`studio.html/js/css`), screenspace (`screenspace.html/js/css`), transcripts (`transcripts.html/js/css`). Shared utilities and constants live in `utils.js` (loaded before each page's main script); `card-scrubber.{js,css}` provides reusable hover-to-scrub on sprite-sheet thumbnails. |
 
 > Subsystem internals (data contracts, key function signatures, merge/dedup rules) live in each module's docstring. The sections below describe how each subsystem is launched and how it fits into the overall flow; for internals, open the referenced module.
 
@@ -41,17 +39,13 @@ Opt-in via `--gallery [VIDEO]` or interactive `gv`/`gallery`. Uses the same inje
 
 Opt-in via `--studio`; requires a spreadsheet. Starts a Flask app via `start_combined_server()` at `config.SERVER_PORT` (8089), serving `assets/web/studio.html`.
 
-## Insights ([insights.py](insights.py), [insights_server.py](insights_server.py))
-
-Opt-in via `--insights`; no spreadsheet required — reads from `clipgen_manifest.json`. Served at `/insights/` by the same Flask server as Studio. The exported `insights_viewer.html` is generated by `finalize_insights_viewer_data()` / `generate_insights_viewer()` in [viewer.py](viewer.py). The frontend source file is `assets/web/insights-builder.{html,js,css}`.
-
 ## Screenspace ([screenspace.py](screenspace.py), [screenspace_server.py](screenspace_server.py))
 
 Opt-in via `--screenspace` or interactive `ss`/`screenspace`; no spreadsheet required. Served at `/screenspace/` by the combined Flask server.
 
 ## Artifact Manifest ([viewer.py](viewer.py))
 
-Opt-in via `--manifest` or `config.MANIFEST_ENABLED`. Writes `clipgen_manifest.json` alongside artifacts. Key functions: `save_manifest`, `load_manifest_artifacts`, `load_manifest_reels`. Consumed by Insights, `--regenerate`, and standalone `--viewer`.
+Opt-in via `--manifest` or `config.MANIFEST_ENABLED`. Writes `clipgen_manifest.json` alongside artifacts. Key functions: `save_manifest`, `load_manifest_artifacts`, `load_manifest_reels`. Consumed by `--regenerate` and standalone `--viewer`.
 
 ## Transcription ([transcripts.py](transcripts.py))
