--- assets/web/insights-builder.html
@@ -1,107 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>clipgen Insights</title>
-  <link rel="icon" type="image/svg+xml" href="logos/favicon.svg">
-  <link rel="icon" type="image/png" sizes="32x32" href="logos/favicon-32.png">
-  <link rel="icon" type="image/png" sizes="192x192" href="logos/favicon-192.png">
-  <link rel="apple-touch-icon" sizes="180x180" href="logos/apple-touch-icon-180.png">
-  <link rel="apple-touch-icon" sizes="167x167" href="logos/apple-touch-icon-167.png">
-  <link rel="apple-touch-icon" sizes="152x152" href="logos/apple-touch-icon-152.png">
-  <link rel="apple-touch-icon" sizes="120x120" href="logos/apple-touch-icon-120.png">
-  <link rel="stylesheet" href="tokens.css">
-  <link rel="stylesheet" href="insights-builder.css">
-  <link rel="stylesheet" href="settings-modal.css">
-</head>
-<body>
-  <header id="builderHeader">
-    <div id="headerTop">
-      <h1><span class="brand-mark" aria-hidden="true"></span>clipgen <span class="frontend-switcher" data-current="insights"><span class="frontend-switcher-trigger" tabindex="0" role="button" aria-haspopup="true" aria-expanded="false"><span class="header-light">Insights</span><span class="frontend-switcher-chevron" aria-hidden="true"></span></span><span class="frontend-switcher-panel" role="menu" aria-hidden="true"><a class="frontend-switcher-item" href="/studio/" data-frontend="studio" role="menuitem">Studio</a><a class="frontend-switcher-item" href="/screenspace/" data-frontend="screenspace" role="menuitem">Screenspace</a><a class="frontend-switcher-item" href="/transcripts/" data-frontend="transcripts" role="menuitem">Transcripts</a></span></span></h1>
-      <div id="headerActions">
-        <span id="unsavedIndicator" class="hidden">Unsaved changes</span>
-        <button id="discardBtn" class="btn" disabled>Discard</button>
-        <button id="saveAllBtn" class="btn btn-primary" disabled>Save</button>
-        <button id="generateViewerBtn" class="btn">Generate Viewer</button>
-        <button id="settingsBtn" type="button" class="btn btn-small btn-icon" title="Settings" aria-label="Settings">
-          <span class="settings-trigger-icon"></span>
-        </button>
-        <button id="themeToggle" type="button" aria-label="Toggle dark mode" aria-pressed="false">
-          <span class="theme-toggle-icon theme-icon-sun"></span>
-          <span class="theme-toggle-icon theme-icon-moon"></span>
-        </button>
-      </div>
-    </div>
-    <div id="headerMeta">
-      <span id="artifactCount">0 artifacts</span>
-      <span class="meta-sep">&middot;</span>
-      <span id="insightCount">0 insights</span>
-    </div>
-  </header>
-
-  <main id="builderLayout">
-    <aside id="mediaLibrary">
-      <div id="sidebarHeader">
-        <h2>Media Library</h2>
-        <button id="collapseBtn" type="button" aria-label="Collapse sidebar" title="Collapse sidebar"><svg width="10" height="10" viewBox="0 0 10 10" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"><path d="M6.5 1.5 3.5 5 6.5 8.5"/></svg></button>
-      </div>
-
-      <div id="filterPanel">
-        <div class="filter-row">
-          <select id="filterParticipant"><option value="">All participants</option></select>
-          <select id="filterCategory"><option value="">All categories</option></select>
-        </div>
-        <div class="filter-row">
-          <select id="filterSeverity"><option value="">All severities</option></select>
-          <div id="typeFilters">
-            <label><input type="checkbox" value="clip" checked> Clip</label>
-            <label><input type="checkbox" value="screen" checked> Screenshot</label>
-            <label><input type="checkbox" value="gif" checked> GIF</label>
-          </div>
-        </div>
-        <input type="text" id="filterSearch" placeholder="Search artifacts..." autocomplete="off">
-      </div>
-
-      <div id="artifactListHeader" class="artifact-list-header">
-        <div id="artifactSortBar" class="artifact-sort-bar" role="toolbar" aria-label="Sort artifact list">
-          <button type="button" class="artifact-sort-btn" data-sort="severity" aria-pressed="false">!<span class="sort-dir" aria-hidden="true"></span></button>
-          <button type="button" class="artifact-sort-btn" data-sort="chrono" aria-pressed="false">&#9201;<span class="sort-dir" aria-hidden="true"></span></button>
-          <button type="button" class="artifact-sort-btn" data-sort="duration" aria-pressed="false">&#8596;<span class="sort-dir" aria-hidden="true"></span></button>
-          <button type="button" class="artifact-sort-btn" data-sort="alpha" aria-pressed="false">A<span class="sort-dir" aria-hidden="true"></span></button>
-        </div>
-      </div>
-      <div id="artifactGrid"></div>
-      <div id="resizeHandle"></div>
-    </aside>
-
-    <section id="insightEditor">
-      <div id="editorToolbar">
-        <button id="newInsightBtn" class="btn btn-primary">+ New Insight</button>
-      </div>
-      <div id="insightCards">
-        <div class="skeleton skeleton-card"></div>
-        <div class="skeleton skeleton-card"></div>
-        <div class="skeleton skeleton-card"></div>
-      </div>
-    </section>
-  </main>
-
-  <div id="addPopover" class="popover hidden">
-    <div id="popoverInsightSelect" class="hidden"></div>
-    <div class="popover-buckets">
-      <button data-bucket="causes" class="popover-btn bucket-causes">Causes</button>
-      <button data-bucket="behaviors" class="popover-btn bucket-behaviors">Behaviors</button>
-      <button data-bucket="impacts" class="popover-btn bucket-impacts">Impacts</button>
-    </div>
-  </div>
-
-  <div id="toast" class="toast hidden"></div>
-
-  <script src="utils.js" defer></script>
-  <script src="dev-token-tweak.js" defer data-dev-only></script>
-  <script src="grid-bg.js" defer></script>
-  <script src="settings-modal.js" defer></script>
-  <script src="insights-builder.js" defer></script>
-</body>
-</html>
