--- assets/web/insights-viewer.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>clipgen Insights</title>
-  <link rel="icon" type="image/svg+xml" href="data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 100' fill='none' stroke='%230a0a0a' stroke-width='11' stroke-linecap='square' stroke-linejoin='miter'><path d='M 18 18 L 82 18 L 82 82'/><path d='M 18 40 L 60 40 L 60 82'/><path d='M 18 62 L 38 62 L 38 82'/></svg>">
-  <link rel="stylesheet" href="insights-viewer.css">
-</head>
-<body>
-  <header id="viewerHeader">
-    <div id="headerTop">
-      <h1 id="studyTitle"><span class="brand-mark" aria-hidden="true"></span>clipgen <span class="header-light">Insights</span></h1>
-      <button id="themeToggle" type="button" aria-label="Toggle dark mode" aria-pressed="false">
-        <span class="theme-toggle-icon theme-icon-sun"></span>
-        <span class="theme-toggle-icon theme-icon-moon"></span>
-      </button>
-    </div>
-    <div id="headerMeta"></div>
-  </header>
-
-  <main id="content">
-    <section id="insightsIndex"></section>
-    <section id="insightDetail" class="hidden"></section>
-  </main>
-
-  <footer id="viewerFooter" class="hidden">
-    <a id="timelineLink" href="#">View all artifacts in Timeline Viewer &rarr;</a>
-  </footer>
-
-  <!-- CLIPGEN_DATA_HERE -->
-  <script src="utils.js" defer></script>
-  <script src="grid-bg.js" defer></script>
-  <script src="insights-viewer.js" defer></script>
-</body>
-</html>
