+++ assets/web/studio.html
@@ -20,7 +20,7 @@
 <body>
   <header id="studioHeader">
     <div id="studioHeaderTop">
-      <h1><span class="brand-mark" aria-hidden="true"></span>clipgen <span class="frontend-switcher" data-current="studio"><span class="frontend-switcher-trigger" tabindex="0" role="button" aria-haspopup="true" aria-expanded="false"><span class="header-light">Studio</span><span class="frontend-switcher-chevron" aria-hidden="true"></span></span><span class="frontend-switcher-panel" role="menu" aria-hidden="true"><a class="frontend-switcher-item" href="/insights/" data-frontend="insights" role="menuitem">Insights</a><a class="frontend-switcher-item" href="/screenspace/" data-frontend="screenspace" role="menuitem">Screenspace</a><a class="frontend-switcher-item" href="/transcripts/" data-frontend="transcripts" role="menuitem">Transcripts</a></span></span></h1>
+      <h1><span class="brand-mark" aria-hidden="true"></span>clipgen <span class="frontend-switcher" data-current="studio"><span class="frontend-switcher-trigger" tabindex="0" role="button" aria-haspopup="true" aria-expanded="false"><span class="header-light">Studio</span><span class="frontend-switcher-chevron" aria-hidden="true"></span></span><span class="frontend-switcher-panel" role="menu" aria-hidden="true"><a class="frontend-switcher-item" href="/screenspace/" data-frontend="screenspace" role="menuitem">Screenspace</a><a class="frontend-switcher-item" href="/transcripts/" data-frontend="transcripts" role="menuitem">Transcripts</a></span></span></h1>
       <div id="headerActions">
         <button id="buildTimelineViewerBtn" class="btn btn-small btn-icon">
           <svg width="14" height="14" viewBox="0 0 16 16" fill="currentColor">
