+++ assets/web/transcripts.html
@@ -18,7 +18,7 @@
 <body>
   <header id="trHeader">
     <div id="headerTop">
-      <h1><span class="brand-mark" aria-hidden="true"></span>clipgen <span class="frontend-switcher" data-current="transcripts"><span class="frontend-switcher-trigger" tabindex="0" role="button" aria-haspopup="true" aria-expanded="false"><span class="header-light">Transcripts</span><span class="frontend-switcher-chevron" aria-hidden="true"></span></span><span class="frontend-switcher-panel" role="menu" aria-hidden="true"><a class="frontend-switcher-item" href="/studio/" data-frontend="studio" role="menuitem">Studio</a><a class="frontend-switcher-item" href="/insights/" data-frontend="insights" role="menuitem">Insights</a><a class="frontend-switcher-item" href="/screenspace/" data-frontend="screenspace" role="menuitem">Screenspace</a></span></span></h1>
+      <h1><span class="brand-mark" aria-hidden="true"></span>clipgen <span class="frontend-switcher" data-current="transcripts"><span class="frontend-switcher-trigger" tabindex="0" role="button" aria-haspopup="true" aria-expanded="false"><span class="header-light">Transcripts</span><span class="frontend-switcher-chevron" aria-hidden="true"></span></span><span class="frontend-switcher-panel" role="menu" aria-hidden="true"><a class="frontend-switcher-item" href="/studio/" data-frontend="studio" role="menuitem">Studio</a><a class="frontend-switcher-item" href="/screenspace/" data-frontend="screenspace" role="menuitem">Screenspace</a></span></span></h1>
       <div id="headerActions">
         <span id="trStatusIndicator" class="status-indicator" tabindex="0" role="status" aria-label="Transcription status"></span>
         <span id="trStatusSr" class="sr-only" aria-live="polite"></span>
