+++ video.py
@@ -1718,107 +1718,3 @@ def generate_interval_captures(
         f"Gallery complete: {len(artifacts)} of {total} captures succeeded."
     )
     return artifacts
-
-
-def generate_sprite_sheet(
-    input_file: str,
-    output_file: str,
-    *,
-    frame_count: int | None = None,
-    thumb_width: int | None = None,
-    min_interval: int | None = None,
-) -> dict[str, Any] | None:
-    """Generate a sprite sheet (contact sheet) from a video file.
-
-    Extracts frames at regular intervals and tiles them into a single PNG.
-    Used for hover-to-scrub previews in the insights builder.
-
-    Returns:
-        Dict with sprite metadata (cols, rows, frameCount, frameWidth,
-        frameHeight, interval) on success, or None on failure.
-    """
-    from math import ceil, sqrt
-
-    frame_count = frame_count or config.SPRITE_SHEET_FRAME_COUNT
-    thumb_width = thumb_width or config.SPRITE_SHEET_THUMB_WIDTH
-    min_interval = min_interval or config.SPRITE_SHEET_MIN_INTERVAL
-
-    if not Path(input_file).is_file():
-        utils.verbose_print(f"Sprite sheet skipped, input not found: '{input_file}'")
-        return None
-
-    duration = get_file_duration(input_file)
-    if duration is None or duration <= 0:
-        return None
-
-    interval = max(min_interval, duration // frame_count)
-    actual_frames = min(frame_count, max(1, duration // interval))
-    if actual_frames <= 0:
-        return None
-
-    cols = ceil(sqrt(actual_frames))
-    rows = ceil(actual_frames / cols)
-
-    # Get source video dimensions for aspect-ratio-correct frame height
-    frame_height = round(thumb_width * 9 / 16)  # fallback to 16:9
-    props = probe_video_properties(input_file)
-    if props and props["width"] > 0:
-        frame_height = round(thumb_width * props["height"] / props["width"])
-
-    if config.DEBUGGING:
-        ic(input_file, output_file, actual_frames, cols, rows, interval)
-        utils.debug_print("Debugging enabled, not calling ffmpeg for sprite sheet.")
-        return {
-            "cols": cols,
-            "rows": rows,
-            "frameCount": actual_frames,
-            "frameWidth": thumb_width,
-            "frameHeight": frame_height,
-            "interval": interval,
-        }
-
-    ffmpeg_command = [
-        "ffmpeg",
-        "-y",
-        "-loglevel",
-        config.FFMPEG_LOGLEVEL,
-        "-i",
-        input_file,
-        "-vf",
-        f"fps=1/{interval},scale={thumb_width}:-1,tile={cols}x{rows}",
-        "-frames:v",
-        "1",
-        output_file,
-    ]
-    utils.debug_print(f"ffmpeg sprite sheet command: {' '.join(ffmpeg_command)}")
-
-    ffmpeg_result = run_ffmpeg_process(
-        ffmpeg_command,
-        input_file=input_file,
-        output_file=output_file,
-        os_error_message="ffmpeg could not run for sprite sheet generation.",
-    )
-    if ffmpeg_result is None:
-        return None
-
-    if ffmpeg_result.returncode != 0:
-        error_details = [f"Input: '{input_file}', Output: '{output_file}'"]
-        utils.error_print(
-            f"ffmpeg sprite sheet failed with exit code {ffmpeg_result.returncode}",
-            _add_ffmpeg_stderr(error_details, ffmpeg_result),
-        )
-        return None
-    if not verify_output_file(output_file, "ffmpeg sprite sheet"):
-        return None
-
-    utils.verbose_print(
-        f"+ Generated sprite sheet '{Path(output_file).name}' ({actual_frames} frames, {cols}x{rows} grid)"
-    )
-    return {
-        "cols": cols,
-        "rows": rows,
-        "frameCount": actual_frames,
-        "frameWidth": thumb_width,
-        "frameHeight": frame_height,
-        "interval": interval,
-    }
